+++ Main/ReplayParser.ReplaySorter/Backup/Models/ReplayBackup.cs
@@ -8,8 +8,8 @@ namespace ReplayParser.ReplaySorter.Backup.Models
 {
     public class ReplayBackup
     {
-        public int BackupId { get; set; }
-        public int ReplayId { get; set; }
+        public long BackupId { get; set; }
+        public long ReplayId { get; set; }
         public string FileName { get; set; }
 
         public Replay Replay { get; set; }
